Refatoração dos serviços de Domínio

+++ Pitang.Kifome.Domain.Entities/User.cs
@@ -10,7 +10,7 @@ namespace Pitang.Kifome.Domain.Entities
         public string Name { get; set; }
         public string Email { get; set; }
         public string Password { get; set; }
-        public int Rate { get; set; }
+        public float Rate { get; set; }
         public IList<Delivery> Deliveries { get; set; }
     }
 }
